rename all the modal components to HKModal
