+++ src/index.ts
@@ -1,4 +1,4 @@
 export { default as HKButton } from './HKButton'
-export { default as Modal } from './Modal'
-export { default as ModalHeader } from './ModalHeader'
-export { default as ModalFooter } from './ModalFooter'
+export { default as HKModal } from './HKModal'
+export { default as HKModalHeader } from './HKModalHeader'
+export { default as HKModalFooter } from './HKModalFooter'
